Pad sections with zero instead of undefined data
The ELF specification dictates that, when `memsz` exceeds `filesz`, the
"extra" bytes should be taken to be zero. Previously, the simulator was
not setting those bytes to zero, instead keeping them at their randomly
initialized values. This commit sets those extra bytes to zero.

Note that everything not specified in the ELF, including the heap,
remains randomly initialized.